+++ elements/rh-card/rh-card.ts
@@ -48,30 +48,33 @@ export class RhCard extends LitElement {
    * Card always resets its context to `base`, unless explicitly provided with a `color-palette`.
    */
   @colorContextProvider()
-  @property({ reflect: true, attribute: 'color-palette' }) colorPalette?: 'darkest' | 'lightest' | 'lighter';
+  @property({ reflect: true, attribute: 'color-palette' })
+    colorPalette?: 'darkest' | 'lightest' | 'lighter';
 
   #slots = new SlotController(this, 'header', null, 'footer');
 
   override render() {
     const { on = '', colorPalette = '' } = this;
     return html`
-     <article part="container" id="container" class="${classMap({ [on]: !!on, [colorPalette]: !!colorPalette })}">
-        <header id="header"
-                part="header"
-                class="${classMap({ empty: !this.#slots.hasSlotted('header') })}">
+     <div id="container"
+          part="container"
+          class="${classMap({ [on]: !!on, [colorPalette]: !!colorPalette })}">
+        <div id="header"
+             part="header"
+             class="${classMap({ empty: !this.#slots.hasSlotted('header') })}">
           <slot name="header"></slot>
-        </header>
+        </div>
         <div id="body"
              part="body"
              class="${classMap({ empty: !this.querySelector(':not([slot])') })}">
           <slot></slot>
         </div>
-        <footer id="footer"
-                part="footer"
-                class="${classMap({ empty: !this.#slots.hasSlotted('footer') })}">
+        <div id="footer"
+             part="footer"
+             class="${classMap({ empty: !this.#slots.hasSlotted('footer') })}">
           <slot name="footer"></slot>
-        </footer>
-      </article>
+        </div>
+      </div>
     `;
   }
 }